Module peng_quad::config

Configuration module This module contains the configuration for the simulation, quadrotor, PID controller, IMU, maze, camera, mesh, and planner schedule. The configuration is loaded from a YAML file using the serde library. The configuration is then used to initialize the simulation, quadrotor, PID controller, IMU, maze, camera, mesh, and planner schedule.

Structs

CameraConfig

Configuration for the camera

Config

Configuration for the simulation

ImuConfig

Configuration for the IMU

MazeConfig

Configuration for the maze

MeshConfig

Configuration for the mesh

PIDControllerConfig

Configuration for the PID controller

PIDGains

Configuration for PID gains

PlannerStep

Configuration for a planner step

QuadrotorConfig

Configuration for the quadrotor

SimulationConfig

Configuration for the simulation