peng_quad
Function log_depth_image

pub fn log_depth_image(
    rec: &RecordingStream,
    depth_image: &[f32],
    resolution: (usize, usize),
    min_depth: f32,
    max_depth: f32,
) -> Result<(), SimulationError>

log depth image data to the rerun recording stream

Arguments

rec - The rerun::RecordingStream instance

depth_image - The depth image data

resolution - The width and height of the depth image

min_depth - The minimum depth value

max_depth - The maximum depth value

Errors

If the data cannot be logged to the recording stream

Example

use peng_quad::log_depth_image;
let rec = rerun::RecordingStreamBuilder::new("log.rerun").connect().unwrap();
let depth_image = vec![0.0; 640 * 480];
log_depth_image(&rec, &depth_image, (640usize, 480usize), 0.0, 1.0).unwrap();