peng_quad

Function log_maze_obstacles

pub fn log_maze_obstacles(
    rec: &RecordingStream,
    maze: &Maze,
) -> Result<(), SimulationError>

Log the maze obstacles to the rerun recording stream

Arguments

rec - The rerun::RecordingStream instance

maze - The maze instance

Errors

If the data cannot be logged to the recording stream

Example

use peng_quad::{Maze, log_maze_obstacles};
let rec = rerun::RecordingStreamBuilder::new("log.rerun").connect().unwrap();
let mut maze = Maze::new([-1.0, -1.0, -1.0], [1.0, 1.0, 1.0], 5, [0.1, 0.1, 0.1], [0.1, 0.5]);
log_maze_obstacles(&rec, &maze).unwrap();